# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{workspaceId}/columns:
    get:
      operationId: getColumns
      tags: [columns]
      summary: List columns
      description: Returns all columns for the given workspace.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      responses:
        "200":
          description: List of columns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Column"
    post:
      operationId: createColumn
      tags: [columns]
      summary: Create a column
      description: Creates a new column in the given workspace.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateColumnRequest"
      responses:
        "201":
          description: Column created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Column"
  /workspaces/{workspaceId}/columns/{columnId}:
    patch:
      operationId: updateColumn
      tags: [columns]
      summary: Update a column
      description: Updates an existing column.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
        - $ref: "#/components/parameters/columnId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateColumnRequest"
      responses:
        "200":
          description: Column updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Column"
    delete:
      operationId: deleteColumn
      tags: [columns]
      summary: Delete a column
      description: Deletes a column and reassigns its tasks to another column.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
        - $ref: "#/components/parameters/columnId"
        - name: reassignToId
          in: query
          required: true
          description: The column ID to reassign tasks to before deleting this column.
          schema:
            type: string
      responses:
        "204":
          description: Column deleted
  /workspaces/{workspaceId}/columns/reorder:
    post:
      operationId: reorderColumns
      tags: [columns]
      summary: Reorder columns
      description: Updates the order of all columns in the workspace.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderColumnsRequest"
      responses:
        "200":
          description: Columns reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Column"
  /workspaces/{workspaceId}/sprints:
    get:
      operationId: getSprints
      tags: [sprints]
      summary: List sprints
      description: Returns all sprints for the given workspace.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      responses:
        "200":
          description: List of sprints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sprint"
    post:
      operationId: createSprint
      tags: [sprints]
      summary: Create a sprint
      description: Creates a new sprint in the given workspace.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSprintRequest"
      responses:
        "201":
          description: Sprint created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sprint"
  /workspaces/{workspaceId}/sprints/{sprintId}:
    patch:
      operationId: updateSprint
      tags: [sprints]
      summary: Update a sprint
      description: Updates an existing sprint.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
        - $ref: "#/components/parameters/sprintId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSprintRequest"
      responses:
        "200":
          description: Sprint updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sprint"
    delete:
      operationId: deleteSprint
      tags: [sprints]
      summary: Delete a sprint
      description: Deletes an existing sprint.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
        - $ref: "#/components/parameters/sprintId"
      responses:
        "204":
          description: Sprint deleted
...
  /workspaces/{workspaceId}/sprint-snapshots:
    get:
      operationId: getSprintSnapshots
      tags: [sprint-snapshots]
      summary: List sprint snapshots
      description: Returns sprint snapshots for the given workspace, optionally filtered by sprint.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
        - name: sprintId
          in: query
          required: false
          description: Filter snapshots by sprint ID.
          schema:
            type: string
      responses:
        "200":
          description: List of sprint snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SprintSnapshot"
    post:
      operationId: upsertSprintSnapshot
      tags: [sprint-snapshots]
      summary: Upsert a sprint snapshot
      description: Creates or updates a sprint snapshot. Unique on sprintId + date.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSprintSnapshotRequest"
      responses:
        "200":
          description: Sprint snapshot upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SprintSnapshot"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    workspaceId:
      name: workspaceId
      in: path
      required: true
      description: The workspace ID.
      schema:
        type: string
    columnId:
      name: columnId
      in: path
      required: true
      description: The column ID.
      schema:
        type: string
    sprintId:
      name: sprintId
      in: path
      required: true
      description: The sprint ID.
      schema:
        type: string
...
  schemas:
...
    Column:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        order:
          type: integer
        semanticStatus:
          type: string
          enum:
            - not-started
            - in-progress
            - done
        color:
          type: string
          nullable: true
        createdAt:
          type: string
          description: ISO datetime
        updatedAt:
          type: string
          description: ISO datetime
      required:
        - id
        - workspaceId
        - name
        - order
        - semanticStatus
        - color
        - createdAt
        - updatedAt
    CreateColumnRequest:
      type: object
      properties:
        name:
          type: string
        order:
          type: integer
        semanticStatus:
          type: string
          enum:
            - not-started
            - in-progress
            - done
        color:
          type: string
          nullable: true
      required:
        - name
        - order
        - semanticStatus
    UpdateColumnRequest:
      type: object
      properties:
        name:
          type: string
        order:
          type: integer
        semanticStatus:
          type: string
          enum:
            - not-started
            - in-progress
            - done
        color:
          type: string
          nullable: true
    ReorderColumnsRequest:
      type: object
      properties:
        columnIds:
          type: array
          items:
            type: string
          description: Ordered list of column IDs representing the new order.
      required:
        - columnIds
    Sprint:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        startDate:
          type: string
          description: YYYY-MM-DD
        endDate:
          type: string
          description: YYYY-MM-DD
        isActive:
          type: boolean
        createdAt:
          type: string
          description: ISO datetime
        updatedAt:
          type: string
          description: ISO datetime
      required:
        - id
        - workspaceId
        - name
        - startDate
        - endDate
        - isActive
        - createdAt
        - updatedAt
    CreateSprintRequest:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: string
          description: YYYY-MM-DD
        endDate:
          type: string
          description: YYYY-MM-DD
        isActive:
          type: boolean
      required:
        - name
        - startDate
        - endDate
    UpdateSprintRequest:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: string
          description: YYYY-MM-DD
        endDate:
          type: string
          description: YYYY-MM-DD
        isActive:
          type: boolean
...
    SprintSnapshot:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        sprintId:
          type: string
        date:
          type: string
          description: YYYY-MM-DD
        total:
          type: integer
        done:
          type: integer
        createdAt:
          type: string
          description: ISO datetime
      required:
        - id
        - workspaceId
        - sprintId
        - date
        - total
        - done
        - createdAt
...
    CreateSprintSnapshotRequest:
      type: object
      properties:
        sprintId:
          type: string
        date:
          type: string
          description: YYYY-MM-DD
        total:
          type: integer
        done:
          type: integer
      required:
        - sprintId
        - date
        - total
        - done